Add a braindump mess enough to find pointerpaths
This is VERY probably not how Cheat Engine does things, as it finds way
more pointers than this method. However, this method is FAST (at least
on the tutorial) even on a single-thread. CE needs to save a "pointermap"
to disk, and then a subsequent pointer-scan uses this pointermap.

Most likely, CE is doing something fancier than just dumping all the
memory, and it is probably comparing these two "fancier dumps" to find
a pointer-path.

However, the approach used in this commit (although really messy),
consists on taking two "snapshots" of the memory, and knowing where
a desired value is located in both.

Then, the pointer value of the address is searched in the second
snapshot (within a certain range, it does not need to be exact).
For every value found, a certain offset will have been used. Now, the
pointer value minus this exact offset must be found exactly on the
other snapshot (it does not matter which snapshot you start with).

If a pointer value in the other snapshot is found, then this offset
is promising. Now the addresses of these pointer values are used to
recurse.

Once the address of one of the pointer values is a base address, there
is no need to keep searching. We have a path of addresses and offsets
we can follow from a base address all the way up to the original value!

This process may generate a tree (multiple addresses may point at some
higher pointer value, maybe with different offsets), so an additional
post-processing step is used to convert the tree into a simpler-to-use
list of lists.

For example, if the address finding yields the following addresses:

    400, 300, 450, 300, 200, 100

(The hundreds' also represent the depth). Then this represents the
following call-stack tree:

       100
        |
       200
        |
       300
       / \
    400   450

These offsets being (100 (100 (100 (100, 150)))) will produce the lists:

    100, 100, 100, 100
    100, 100, 100, 150