Add a second implementation of PathFinder
This version uses more memory, but can easily be modified to run in a
multi-threaded environment.

Instead of using the variables in the call-stack to save the current
path being scanned, the data-structure is separated into a "linked list"
of nodes walked (which could be shared between threads).

Instead of recursing, work is added to a different list (used as a
queue), and the program loops until it's empty.

When reaching a base address, the index of the "good node" is saved.
After the loop finishes, these "good nodes" can be re-visited, and the
parent chain followed.

This also makes it easier to change the algorithm from depth-first
to breadth-first.