Copy pathfinder to make an optimized version
SnapshotOpt is like Snapshot, but contains additional information,
which lets us avoid scanning a lot of blocks.

For example, rather than scanning 95 blocks, we scan an average of
3.145 blocks (median 2, SD 6.12), which greatly reduces the amount
of work that needs to be done.

Snapshot can be seen as a special-case of SnapshotOpt where "blocks
to scan" is "all of them".