Greatly improve snapshot optimizer
Rather than scanning ALL the pointer values (potentially several
gigabytes) which point into the CURRENT block, scan the pointer
values of the CURRENT block and check into which block they point to.

Assuming "check into which block they point to" is O(log n)
(such as when using binary search), this brings the runtime from
O(n²) to O(n log n).

When the number of pointer-values is as high as 750M (~6GB),
and the number of blocks is over a thousand (such as 3.5K), this
reduces the amount of work from 2625G to 8.8G (nearly 300 times
faster).

In addition to that, the iterator can be simpler, as it no longer
needs to embed the real address in order to determine which block
the pointer-value came from. This should also improve the code
generation for a more modest boost.